requirements.txt
sentence-transformers>=2.2.2
faiss-cpu>=1.7.2
numpy
scikit-learn
pdfplumber
transformers
torch
tqdm
